Done Refactoring's
Poor implementation of HelpCommand. You must ensure that adding a new command involves adding
a new subclass of Command and adding one line to the AVAILABLE_COMMANDS. NOTHING ELSE! It
should not involve any looking for other arrays or switches that need extending in classes such
as the HelpCommand class.
One solution: define a static method of the Command class that builds a string by looping through
the available commands adding:
  details + ": " + help       (what did you think these attributes of Command were for?)
plus a newline instruction for each one. Then in the execute method of the HelpCommand call this
static method and print the string returned.

execute method of several commands: instead of multiple method calls, e.g. MoveDownCommand:
  result = game.movePlayer(result, direction);
  game.incrementCyle();
there should be just one call; in the case of this example, you should call incrementCycle and
from inside playerN. The less the command classes know about how the game works, the better!

Passing the symbol up to GameElement via the constructor and then putting getSymbol in GameElement
- very good except for the fact that getSymbol should be called toString (what do you think the
toString method is for?)

Why do game elements need methods to set their coordinates? (they don't!)

**** BREAKING ENCAPSULATION ****
getObjectInPosition is breaking encapsulation, as is calling the ArrayList get method of the
container from Game. It wouldn't be breaking encapsulation if the type of the object
returned were Collider. If you break encapsulation like this in the second assignment,
you will fail. There is no excuse for doing this in the second assignment. Please read the
problem statement and the 6th generic comment to the first assignment and fix the problem.

For the 'event handling' methods onDelete, onEnter,... instead of declaring an abstract
method in GameElement and then overwriting it with an empty method in many GameElement subclasses,
it would be better to define a concrete but empty method in GameElement and then only overwrite
it in those GameElement subclasses where it is not empty (or move the empty method up to
the interface as a default method).