Simon Pickin
14:36 (4 hours ago)

to me

I have checked out your code.

Firstly and most importantly, in the container you have the comment
   // This could be breaking encapsulation
Indeed it is. YOU MUST REMOVE IT. If you loop through someone else's
data, you are likely to break encapsulation. There is no need to do it.

Next, in the game, you have a comment.
   //@Simon you told us to move it to the Game in your eMail from the 29.11.
I cannot find any such statement in an e-mail of mine. This code should
be in the execute of the WaveAction. The limits for the two loops in the
execute of the WaveAction are the visibility and the road width; to
obtain them, it can invoke methods of the game (it has a reference to
the game).

Going back to the original question, there is not need to export the
size from the container, even if you do it while respecting information
hiding. There should be no reason to have any loops (using size) in
Game. Move two of them to the container and the other to the WaveAction.

     Regards